:hammer: Move metadata/link into models + comment
