+++ examples/write_link.rs
@@ -1,7 +1,8 @@
-use in_toto::metadata::{LinkMetadataBuilder, VirtualTargetPath};
+use in_toto::models::link::metadata::{LinkMetadataBuilder};
 use in_toto::crypto::{PrivateKey, KeyType, SignatureScheme};
 use in_toto::interchange::Json;
 use serde_json;
+use in_toto::models::helpers::{VirtualTargetPath};
 
 fn main() {
     // Generate a new Ed25519 signing key
