+++ src/models/layout.rs
@@ -1,0 +1,1 @@
+//! in-toto layout: used by the project owner to generate a desired supply chain layout file.
