+++ src/models/mod.rs
@@ -1,0 +1,6 @@
+//! Models used in in-toto
+
+pub mod link;
+pub mod metadata;
+pub mod layout;
+pub mod helpers;
