+++ src/verification.rs
@@ -1,0 +1,1 @@
+//! A tool to be used by the client to perform verification on the final product.
